Fixed modelardb_common compile errors after merge and moved argument functions back
