+++ crates/modelardb_common/src/arguments.rs
@@ -149,6 +149,38 @@ pub fn decode_argument(data: &[u8]) -> Result<(&str, &[u8]), Status> {
     Ok((argument, remaining_bytes))
 }
 
+/// Parse the arguments in `data` and return the arguments to connect to an
+/// [`Amazon S3`](object_store::aws::AmazonS3) object store and what is remaining of `data`
+/// after parsing. If `data` is missing arguments, [`Status`] is returned.
+pub async fn extract_s3_arguments(data: &[u8]) -> Result<(&str, &str, &str, &str, &[u8]), Status> {
+    let (endpoint, offset_data) = decode_argument(data)?;
+    let (bucket_name, offset_data) = decode_argument(offset_data)?;
+    let (access_key_id, offset_data) = decode_argument(offset_data)?;
+    let (secret_access_key, offset_data) = decode_argument(offset_data)?;
+
+    Ok((
+        endpoint,
+        bucket_name,
+        access_key_id,
+        secret_access_key,
+        offset_data,
+    ))
+}
+
+/// Parse the arguments in `data` and return the arguments to connect to an
+/// [`Azure Blob Storage`](object_store::azure::MicrosoftAzure)
+/// object store and what is remaining of `data` after parsing. If `data` is missing arguments,
+/// [`Status`] is returned.
+pub async fn extract_azure_blob_storage_arguments(
+    data: &[u8],
+) -> Result<(&str, &str, &str, &[u8]), Status> {
+    let (account, offset_data) = decode_argument(data)?;
+    let (access_key, offset_data) = decode_argument(offset_data)?;
+    let (container_name, offset_data) = decode_argument(offset_data)?;
+
+    Ok((account, access_key, container_name, offset_data))
+}
+
 #[cfg(test)]
 mod test {
     use super::*;
