+++ crates/modelardb_common/src/metadata/table_metadata_manager.rs
@@ -33,8 +33,8 @@ use arrow_flight::{IpcMessage, SchemaAsIpc};
 use dashmap::DashMap;
 use datafusion::common::{DFSchema, ToDFSchema};
 use datafusion::prelude::col;
-use deltalake::kernel::{DataType, StructField};
-use deltalake::DeltaTableError;
+use deltalake_core::kernel::{DataType, StructField};
+use deltalake_core::DeltaTableError;
 use object_store::path::Path;
 use object_store::ObjectMeta;
