+++ crates/modelardb_manager/Cargo.toml
@@ -27,7 +27,7 @@ path = "src/main.rs"
 arrow.workspace = true
 arrow-flight.workspace = true
 chrono.workspace = true
-deltalake = { workspace = true, features = ["datafusion", "s3"] }
+deltalake-core.workspace = true
 futures.workspace = true
 modelardb_common = { path = "../modelardb_common" }
 object_store = { workspace = true, features = ["aws", "azure"] }
